Enforce no warnings for MSC.

+++ gip/operation/Interpolate.h
@@ -156,14 +156,14 @@ namespace gip {
       
       const Pixel* p = elements + y0 * dimension.getWidth() + x0;
       
-      if ((x0 >= 0) && (x0 < dimension.getWidth())) {
-        if ((y0 >= 0) && (y0 < dimension.getHeight())) {
+      if ((x0 >= 0) && (static_cast<unsigned int>(x0) < dimension.getWidth())) {
+        if ((y0 >= 0) && (static_cast<unsigned int>(y0) < dimension.getHeight())) {
           const long double w0 = (1 - xFraction) * (1 - yFraction);
           red += w0 * p[0].red;
           green += w0 * p[0].green;
           blue += w0 * p[0].blue;
         }
-        if (((y0 + 1) >= 0) && ((y0 + 1) < dimension.getHeight())) {
+        if (((y0 + 1) >= 0) && (static_cast<unsigned int>(y0 + 1) < dimension.getHeight())) {
           const long double w2 = (1 - xFraction) * yFraction;
           red += w2 * p[dimension.getWidth()].red;
           green += w2 * p[dimension.getWidth()].green;
@@ -173,14 +173,14 @@ namespace gip {
       
       ++p;
       
-      if (((x0 + 1) >= 0) && ((x0 + 1) < dimension.getWidth())) {
-        if ((y0 >= 0) && (y0 < dimension.getHeight())) {
+      if (((x0 + 1) >= 0) && (static_cast<unsigned int>(x0 + 1) < dimension.getWidth())) {
+        if ((y0 >= 0) && (static_cast<unsigned int>(y0) < dimension.getHeight())) {
           const long double w1 = xFraction * (1 - yFraction);
           red += w1 * p[0].red;
           green += w1 * p[0].green;
           blue += w1 * p[0].blue;
         }
-        if (((y0 + 1) >= 0) && ((y0 + 1) < dimension.getHeight())) {
+        if (((y0 + 1) >= 0) && (static_cast<unsigned int>(y0 + 1) < dimension.getHeight())) {
           const long double w3 = xFraction * yFraction;
           red += w3 * p[dimension.getWidth()].red;
           green += w3 * p[dimension.getWidth()].green;
